requirements.txt
Flask
Pillow
email-validator
moviepy
miniupnpc
psutil
pystray
webvtt